pyproject.toml
[tool.poetry]
name = "playlist-converter"
version = "0.1.0"
description = ""
authors = ["Dylan Rochon-Terry <dylan@ecovative.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
async-timeout = "4.0.2"
certifi = "2023.5.7"
charset-normalizer = "2.0.12"
idna = "3.4"
importlib-metadata = "4.8.3"
packaging = "21.3"
pyparsing = "3.0.7"
redis = "4.3.6"
requests = "2.27.1"
six = "1.16.0"
spotipy = "2.23.0"
typing-extensions = "4.1.1"
urllib3 = "1.26.16"
zipp = "3.6.0"
pydub = "^0.25.1"
youtube-search = "^2.1.2"
pytubefix = "^8.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
